backend/requirements.txt
altair==5.5.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
django-cors-headers
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.6
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
narwhals==2.5.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
qrcode==8.2
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
social-auth-app-django==5.5.1
social-auth-core==4.7.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
streamlit==1.50.0
streamlit-cookies-manager==0.2.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xgboost==3.0.5

